+++ client/src/main/java/replicant/AreaOfInterestService.java
@@ -8,7 +8,6 @@ import arez.annotations.Feature;
 import arez.annotations.Observable;
 import arez.annotations.ObservableValueRef;
 import arez.annotations.PreDispose;
-import arez.component.CollectionsUtil;
 import arez.component.DisposeNotifier;
 import java.util.HashSet;
 import java.util.List;