+++ client/src/main/java/replicant/EntityService.java
@@ -6,7 +6,6 @@ import arez.annotations.ArezComponent;
 import arez.annotations.Feature;
 import arez.annotations.Observable;
 import arez.annotations.ObservableValueRef;
-import arez.component.CollectionsUtil;
 import arez.component.ComponentObservable;
 import arez.component.DisposeNotifier;
 import java.util.Collection;