+++ client/src/main/java/replicant/SchemaService.java
@@ -1,6 +1,5 @@
 package replicant;
 
-import arez.component.CollectionsUtil;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;